bug(#507): proper places
